+++ R/dbreg-package.R
@@ -1,4 +1,8 @@
 #' @keywords internal
 "_PACKAGE"
 
-NULL
+# Silence R CMD check NOTEs for NSE
+utils::globalVariables(c(
+ "bin", "x_mid", "u", "u2",      # dbbinsreg
+ "V_beta", "cb_lwr", "cb_upr"    # dbbinsreg / plot
+))