+++ R/dbreg.R
@@ -1,4 +1,4 @@
-#' Run a regression on a database backend.
+#' Run a regression on a database backend
 #'
 #' @md
 #' @description
@@ -211,11 +211,11 @@
 #' @seealso \code{\link[DBI]{dbConnect}} for creating database connections,
 #' \code{\link[duckdb]{duckdb}} for DuckDB-specific connections
 #'
-#' @importFrom DBI dbConnect dbDisconnect dbGetInfo dbGetQuery
-#' @importFrom duckdb duckdb duckdb_register
+#' @importFrom DBI dbConnect dbDisconnect dbExecute dbGetInfo dbGetQuery dbIsValid dbRemoveTable dbWriteTable
+#' @importFrom duckdb duckdb duckdb_register duckdb_unregister
 #' @importFrom Formula Formula
 #' @importFrom Matrix chol2inv crossprod Diagonal sparse.model.matrix
-#' @importFrom stats formula pt reformulate setNames
+#' @importFrom stats aggregate as.formula formula pt reformulate setNames
 #' @importFrom glue glue glue_sql
 #'
 #' @examples
@@ -406,8 +406,10 @@ process_dbreg_inputs = function(
     if (!inherits(data, "data.frame")) {
       stop("`data` must be data.frame.")
     }
-    duckdb_register(conn, "tmp_table_dbreg", data)
-    from_statement = "FROM tmp_table_dbreg"
+    temp_name = sprintf("tmp_table_dbreg_%s", 
+                       gsub("[^0-9]", "", format(Sys.time(), "%Y%m%d_%H%M%S_%OS3")))
+    duckdb_register(conn, temp_name, data)
+    from_statement = paste("FROM", temp_name)
   } else if (!is.null(path)) {
     if (!is.character(path)) {
       stop("`path` must be character.")